requirements.txt
backports.entry-points-selectable==1.1.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
dataclasses==0.6
distlib==0.3.3
easyocr==1.4.1
filelock==3.3.1
future==0.18.2
google-api-core==2.2.0
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
imageio==2.9.0
imutils==0.5.4
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.3
oauthlib==3.1.1
opencv-python-headless==4.5.4.58
Pillow==8.2.0
platformdirs==2.4.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
soupsieve==2.2.1
tifffile==2021.10.12
torch==1.10.0+cu102
torchaudio==0.10.0+cu102
torchvision==0.11.1+cu102
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.9.0
